tty: Unify tty serial ops
