+++ src/drivers/tty/serial/Mybuild
@@ -7,7 +7,6 @@ module serial_dvfs {
 	source "ttys_processing.c"
 	source "ttys_term.c"
 	source "ttys_dvfs.c"
-	source "idesc_serial_dvfs.c"
 	source "idesc_serial.c"
 
 	@IncludeExport(path="drivers", target_name="ttys.h")
@@ -24,7 +23,6 @@ module serial_oldfs {
 	source "ttys_processing.c"
 	source "ttys_term.c"
 	source "ttys_oldfs.c"
-	source "idesc_serial_oldfs.c"
 	source "idesc_serial.c"
 
 	@IncludeExport(path="drivers", target_name="ttys.h")
