+++ src/drivers/tty/serial/idesc_serial.h
@@ -16,6 +16,4 @@ extern struct idesc *idesc_serial_create(struct uart *uart, int __oflags);
 
 extern const struct idesc_ops *idesc_serial_get_ops(void);
 
-extern void idesc_serial_close(struct idesc *idesc);
-
 #endif /* IDESC_SERIAL_H_ */
